[artf6621]: add testing for template support
A few improvements to the UI implemented while writing tests.

* grails-app/i18n/messages.properties
  New resources for field validation

* grails-app/services/com/collabnet/svnedge/console/RepoTemplateService.groovy
  cleaned up some imports and service DI fields that are not used.

  (saveTemplate): modified so view layer can display field errors from attempting to save
  the domain object.

* grails-app/controllers/com/collabnet/svnedge/controller/RepoTemplateController.groovy
  (save): Reflect changes to service
  (handleFileUpload): Add field error, if the file is empty, or does not match mime-types:
  application/octet-stream or application/zip.

* grails-app/views/repoTemplate/create.gsp
  We are showing field level errors, near the form field, so do not need to duplicate these
  errors at the top.  Also don't want the warning/error icon displaying on the field itself,
  we show it just below the field.

* grails-app/views/repo/_leftNav.gsp
  Fixed some of the menu highlighting to better account for the multiple controllers,
  preventing situations where 2 items are highlighted.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2629 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ grails-app/services/com/collabnet/svnedge/console/RepoTemplateService.groovy
@@ -17,22 +17,12 @@
  */
 package com.collabnet.svnedge.console
 
-import java.io.File;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
-import org.apache.commons.compress.archivers.zip.ZipArchiveEntry
-import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream
-import org.hyperic.sigar.FileInfo
+import java.util.zip.ZipFile;
 
 import com.collabnet.svnedge.util.ConfigUtil;
 import com.collabnet.svnedge.domain.RepoTemplate
 
 class RepoTemplateService extends AbstractSvnEdgeService {
-
-    // service dependencies
-    def operatingSystemService
-    def lifecycleService
-    def commandLineService
 
     File getTemplateDirectory() {
         File templateDir = new File(ConfigUtil.dataDirPath(), 'repoTemplates')
@@ -54,7 +44,7 @@ class RepoTemplateService extends AbstractSvnEdgeService {
 
         boolean isArchive = templateFile.name.endsWith(".zip")
         if (isArchive) {
-            def zipFile = new java.util.zip.ZipFile(templateFile)
+            def zipFile = new ZipFile(templateFile)
             template.dumpFile = (zipFile.size() == 1)
     
         } else {
@@ -64,8 +54,9 @@ class RepoTemplateService extends AbstractSvnEdgeService {
         def currentTemplates = RepoTemplate.list()
         if (isInsert) {
             template.displayOrder = 1
+            int i = 2
             for (RepoTemplate t in currentTemplates) {
-                t.displayOrder++
+                t.displayOrder = i++
                 t.save()
             }
         } else {
@@ -75,25 +66,26 @@ class RepoTemplateService extends AbstractSvnEdgeService {
         // give a temp name to allow saving the record
         template.location = "temp/" + templateFile.name
         template.active = true
-        template.save()
-        
-        String filename = (template.dumpFile ? "dump" : "repoArchive") +
-            template.id
-        if (isArchive) {
-            filename += ".zip"
+        boolean success = template.save()
+        if (success) {
+            String filename = (template.dumpFile ? "dump" : "repoArchive") +
+                    template.id
+            if (isArchive) {
+                filename += ".zip"
+            }
+            File finalFile = new File(getTemplateDirectory(), filename)
+            if (finalFile.exists()) {
+                finalFile.delete()
+            }
+            if (!templateFile.renameTo(finalFile)) {
+                finalFile.withOutputStream { out -> 
+                    templateFile.withInputStream { instrm -> out << instrm } }
+                templateFile.delete()
+            }
+            template.location = finalFile.name
+            success = template.save()
         }
-        File finalFile = new File(getTemplateDirectory(), filename)
-        if (finalFile.exists()) {
-            finalFile.delete()
-        }
-        if (!templateFile.renameTo(finalFile)) {
-            finalFile.withOutputStream { out -> 
-                templateFile.withInputStream { instrm -> out << instrm } }
-            templateFile.delete()
-        }
-        template.location = finalFile.name
-        template.save()
-        return true
+        return success
     }
         
     /**
